Added WithdrawalNoticeDto and its Mapper- both are working - Just need to fix the null val on current balance

+++ src/main/java/com/enviro/assessment/grad001/abongilebilly/PaperlessWithdrawals/models/Product.java
@@ -25,7 +25,7 @@ public class Product {
 
     private String name;
 
-    private Long currentBalance;
+    private Integer currentBalance;
 
     @JsonBackReference
     @ManyToOne
